[refactor] 응답 로직 수정
