+++ backend/src/main/java/com/hoppy/app/response/service/ResponseService.java
@@ -1,6 +1,5 @@
 package com.hoppy.app.response.service;
 
-import com.hoppy.app.response.code.SuccessCode;
 import com.hoppy.app.response.dto.Data;
 import com.hoppy.app.response.dto.ResponseDto;
 import org.springframework.http.ResponseEntity;
